Move configuration out, but ensure it's included
We can't use the defaults/ mechanism to avoid the clumsy key_name vs.
ec2_key_name recursion guard, because this isn't a role. It would be
nice, but that alone isn't worth the conversion.

The asserts (in particular, the "all instances in different regions"
one) require the 'onattr' custom filter committed earlier.

(This used to work before the reorganisation, but hasn't been tested
since then.)