Split repmgr role into common/primary/standby roles
This makes it possible to avoid deeply-nested blocks inside roles.

In theory, we could list repmgr_common as a role dependency in
repmgr_{primary,standby}/meta/main.yml, but that (a) causes the role to
be included twice (i.e. it's run for one host and skipped for two, then
run for the two and skipped for the first one), which is slow; and (b)
for some reason the tasks were skipped on the standby. So we do it the
ugly and simple way.