Retag provisioning playbook to use tags consistent with deprovisioning
Also reorganise the validation tasks into a separate block for clarity.

Note that the one-instance-per-region restriction is removed here, but
the rest of the playbook hasn't been updated to suit. I'm committing
this separately just to keep the pieces manageable.

Also note that we always run the ec2 task because the ec2_eip task
depends on its output. That should also be fixed eventually.