Note that deprovision.yml doesn't delete extra volumes
Unfortunately, ec2.py doesn't give us volume_ids in hostvars, so there
is no easy way to discover and remove them.

The solution may be to run «aws ec2 describe-volumes» with the correct
filters and parse the output; we may also want to create the volumes in
a separate step using ec2_vol, tag them properly, and associate them in
another step with the provisioned instances (there's no way to tag the
volumes when they are created using ec2; and this approach would make
it possible to initialise instances from snapshots as well).

For now, however, we must remember to delete extra volumes by hand.