Unroll the /etc/hosts loop as well
If we use a template that does the loop locally, we again avoid making
n² ssh connections.

If we had to cope with other stuff in /etc/hosts, we'd have to do this
in two steps: template the part we know, and assemble it with the part
we don't. But we don't care for the moment.