Disable template caching to fix an elusive bug
We were doing something like this:

    with_nested_dependents:
        - regions
        - group[cluster_tag]|intersect(groups[item.0])|…|list

…and it so happened that groups[regions[0]] was empty (although
groups[regions[1]] was not), but the entire loop was being skipped.
After a lot of debugging, it turned out that the expression was not
being re-evaluated for the new value of item.0 (== regions[1]).

listify_lookup_plugin_terms() offers no way to turn off the caching, so
we do what it does directly instead of calling it. There are suggestions
that the caching shouldn't operate in this situation (where we evaluate
the same expression with different variables), but I haven't bothered to
investigate further, because it's clearly not working.

(While investigating I also discovered a minor problem in the upstream
code, and opened https://github.com/ansible/ansible/pull/13737)