Set ansible_user in the inventory
If config.yml defines ec2_ami_user, it'll be set as the ansible_user for
all hosts in the cluster. So if you're deploying a Debian cluster, you'd
use admin (the default), but you could set ec2-user instead if you were
using the RHEL images.

This can be extended to other attributes (e.g. distribution, to remove
the need to run setup). But that can happen by and by; ansible_user is
the one thing we need to be able to set right now.

Of course, we should be able to override this on a per-host basis, but
that just doesn't seem terribly important to do right now. It's also an
easy task for someone else to pick up later: just check if the instance
has some variable set, and if so, write the variable on the host line.