1) A new repmgrd file will now be copied into /etc/default/repmgrd
2) Introduce "logfile" and "pg_bindir" paramters in repmgr.conf. Without
"pg_bindir" the pg_ctl binary path is not available at all times

3) Changes to the primary/standby repmgr roles to start repmgrd

4) The "standby clone" operation copies conf files from primary to inside
the DATA directory on the standby nodes. Copy them over to the
appropriate locations on the standby. The "repmgr_funcs" shared library
was not getting loaded if the standby used the default old conf files,
which is very wrong