Rework the handling of credentials to use instance profiles
Instance profiles are the correct solution to the various problems noted
with the earlier approach (of placing credentials in user-data):

http://docs.aws.amazon.com/IAM/latest/UserGuide/id_roles_use_switch-role-ec2_instance-profiles.html

This way, we create a role and assign it any permissions that we want,
and when instances are provisioned, they're assigned that profile, and
ephemeral credentials are magically made available to instances when
something (e.g., boto) looks for them.

We retain an empty user-data.j2 template, because we might want to use
it for something in the future.