1) Change the TPA config.yml to use a SINGLE region for the cluster. There's no point really in having standby instances spread across the globe..
2) Add a "variables" section in the end of TPA config.yml. All of these
key/value pairs will be added to the inventory after provisioning making
them visible as "vars" to the subsequent deploy activity. Support in the
AWS provision script was already committed earlier for this.

3) A host of changes in the TPA deploy.yml file to support various with/
without barman configurations. So all the below are possible now:
  i) A TPA cluster with 1 master + 2 standby with repmgr
 ii) A TPA cluster with 1 master + 2 standby with repmgr + 1 exclusive barman
iii) A TPA cluster with 1 master + 2 standby with repmgr + Shared barman

  We have added a second play in deploy.yml which gets the barman host as its
  inventory dynamically and then works on it along with the IP address of the
  host to backup from

4) Introduce "customerid" and "serviceid" variables. These can be passed on
from the TPA GUI component while provisioning. These help in uniquely
identifying the instance from a specific TPA cluster to a shared barman for
configuring backup. These two values are prepended to the name that's given
to the backup config for a specific instance to uniquely identify it amongst
possibly many. The form is basically:
 {{customerid + '_' + serviceid + '_' + hostvars[item]['ec2_tag_db'] + hostvars[item]['ec2_tag_node']

5) A host of changes in barman roles files to accomodate using an already
existing barman instance. Fixing a host of commands to make them use the above
naming convention format.

6) Some pg_hba.conf changes to explicitly add repmgr instances IPs. This is due to
the removal of the earlier catchall '0.0.0.0' type of entries.