Overhaul common role
This role is meant to (a) detect the distribution, (b) do whatever
distribution-specific tasks are required to get a working Python
installation, (c) install system packages.

So we have: distribution.yml, os/${ansible_distribution}.yml,
python.yml, and packages.yml.

We prefer includes based on ansible_distribution in order to avoid a
proliferation of complex logic and "when: ansible_distribution == 'x'"
statements throughout the playbooks.

Note that we switch from 'debian' and 'redhat' to 'Debian', 'RedHat',
and 'Ubuntu', in order to stay compatible with what the built-in setup
module will set ansible_distribution to. This will make it easier to use
setup in hybrid clusters and other situations where we don't know in
advance exactly what we will encounter.