Fix distribution detection code
The problem was that the redhat-release task overwrote "distribution"
even when the task was skipped, but this fix goes further and runs
lsb_release only once (with -ics) and revises the logic suitably.

Also, don't run with sudo because otherwise we ended up setting
ansible_distribution to "sudo: unable to resolve host ip-10-19-1-5".
Oops.