Fix nested block logic
The way nested blocks work in ansible is that the conditions on the
outer and inner blocks are propagated to each nested task and evaluated
at the task level (and not at the block level). So when the first task
in the inner block set ansible_distribution, the condition on the next
task ("ansible_distribution is not defined AND lsb_release|succeeded")
failed, and the task was skipped.

(This problem doesn't affect the RedHat detection block, because the two
nested blocks will never be run together.)

But set_fact can set multiple facts, so we don't need the block at all.