Add a sys/tune role
The purpose of this role is to calculate any values dependent on the
instance's configuration. Right now, we calculate ansible_memtotal_mb
and derive shared_buffers_mb (both set as facts). This replaces the
repeated calculation of memtotal elsewhere. Both these values (and the
ratio used to derive the latter) can be overriden explicitly when
including the role, e.g.:

    - role: sys/tune
      shared_buffers_ratio: 0.15

Any similar calculations should go into this role.

Thanks to agaffney on #ansible-devel for his help to come up with the
extraordinary expression to set an integer fact from an expression.