Overhaul sys/sysctl role
The primary change is that this is definition-driven, with defaults all
in one place, but the caller able to augment or override them at need.
The code is considerably shotrer as a consequence. Another is to avoid
the "if (A-enabled) {disable-A}" pattern; "disable A" suffices.

One particularly nasty problem discovered during this process is that if
you write vm.dirty_bytes (say, 4GB) and then set vm.dirty_ratio to 0 for
clarity, the effective value is 0. Oops.

TODO:

1. We still need to modify transparent huge page behaviour.
2. dirty_bytes=4GB isn't appropriate for small instances; that's just a
   matter of tuning the formula until it makes sense.