Update EPEL release RPM
In theory, we should be using the yum_repository module to install yum
repositories, so that it will find the latest version of the RPM on its
own, but my attempt to do this for the EPEL repo wasn't successful—the
problem being that the (randomly-selected) mirror didn't have the right
repodata, and download.fedoraproject.org does the mirror selection, so I
didn't know what mirrors to specify or how.

We should clearly make it work that way in the long term, but this is
good enough for the moment.