Tag tasks inside common role
The common role performs distribution detection, which subsequent roles
depend on, so we can't just do:

    - role: common
      tags: common

and run the playbook with --skip-tags common.

Instead, we leave the role untagged, but tag the tasks inside it, so
that distribution detection always happens (unless overriden), but the
other tasks can indeed be skipped with --skip-tags common.