Add link to 2ndQPostgres release RPM
It's possible to get yum_repository to work, but not very convincingly.
Despite the problem of linking to the latest version of the release RPM,
it's probably better to stick to it for now. Part of the advantage is
that it'll install the GPG key locally, and set up all the variants of
the repository (2ndQPostgres, 2ndQPostgres-testing).