Create sys/ssh role to enable password-less ssh
This is an improved version of pkg++'s tasks/ssh.yml (for example, it
doesn't use ssh-keyscan on single hosts in a with_items loop, which is
inordinately painful with large clusters).

This code shouldn't be looking at ec2_* values, and it should also be
using cluster_hostname (but we can't, because /etc/hosts isn't updated
yet). Those have to be dealt with at a higher level.

This code will also replace the equivalent in the XL roles.