Generate instance host keys locally
Right now, we blindly accept whatever host keys are auto-generated by
each provisioned instance (via ssh-keyscan). This commit is the first
step towards fixing this problem once and for all.

We now generate RSA and ECDSA keypairs locally and upload them to an S3
bucket (to which the instance profile is granted access); a user-data
script installs them in /etc/ssh on the first boot.

A few notes:

(1) The ansible s3 module returns http URLs with a specified expiration
time to access objects. Unfortunately, you can't just use this URL with
curl because S3 likes to issue responses like this:

<?xml version="1.0" encoding="UTF-8"?>
<Error><Code>TemporaryRedirect</Code><Message>Please re-send this request to the specified temporary endpoint. Continue to use the original request endpoint for future requests.</Message><Bucket>2ndquadrant-tpa-tpa-rhel</Bucket><Endpoint>2ndquadrant-tpa-tpa-rhel.s3-eu-west-1.amazonaws.com</Endpoint><RequestId>20A682984884F3F3</RequestId><HostId>a63fgKtXjHQcgQPwW+QZ81X1QJJFEyhxM+kvzwOVWMyrx+qrHVfIlsdOZ6NPV9cmSu+Epp1jG2A=</HostId></Error>

Of course, curl has no idea what to do with that response. Also, the AWS
cli doesn't have any way to download S3 objects via such HTTP URLs (that
I could find). So we install the aws cli using the .zip bundle, and use
its "s3api get-object" command. That works fine.

(2) Note that we replace (overwrite) the RSA/ECDSA keys, and delete
everything else. We don't try to remove HostKey lines from sshd_config,
though, because sshd will just ignore host keys it can't load. sshd does
not appear to need a restart when replacing host keys either.

(3) We don't have access to ansible_distribution etc. facts during the
provisioning process, so user-data.j2 needs to be written in a manner
that will run on both RedHat and Debian/Ubuntu.

(4) There's a problem with this: the user-data script can take up to a
minute to run. During that time, sshd is accepting connections with the
old host keys. So you can login, and the keys will be replaced, and the
next time you try to connect, you'll get a key verification error. It's
possible that the best way to deal with this is to switch ssh ports, and
wait for sshd to start listening on the new one (which can be made to
happen only after the new keys are installed).

Also, creating our own AMIs with unzip/aws-cli etc. installed already
will also reduce the extra time taken on first boot (but that isn't a
solution in itself; there will always be a race condition unless we
sequence the operations as described above).

The next step is to generate a UserKnownHostsFile locally (for ssh to
use when invoked via ansible), and to replace the multiple ssh-keyscan
invocations on the remote node to identify other hosts in the cluster.
We have all the public keys, so we can generate the files ourselves and
install them in /etc/ssh on every node.