Generate a local known_hosts file for the cluster hosts
This is fairly straightforward: we have the public IP addresses and the
host keys, so we stick them into a file and point ssh at it as the
UserKnownHostsFile.

A few notes:

(1) We don't bother to hash hostnames in known_hosts. It could be done
with a bit of extra effort and some lookup('pipe', 'openssl…') calls,
but it doesn't seem worth it.

(2) The inventory file must contain some path to the known_hosts file.
This brings in new complications, no matter how we do it:

    (a) If we set "{{ cluster_dir }}/known_hosts", the inventory will
    contain absolute paths like /home/ams/…/clusters/test/rhel, which
    are unlikely to work for anyone but ams.

    (b) If we set "known_hosts", then we'll have to make sure that ssh
    is invoked from $cluster_dir, which means we'll have to cd to that
    dir (and we can't just say "~/…/utils/deploy test/rhel" wherever we
    may be, as we could before).

    (c) Likewise, if we set "clusters/$cluster/known_hosts", we'll need
    to be in the top-level directory.

The compromise I settled on is this:

    (a) We pass «-e ansible_ssh_common_args» on the command-line from
    the utils/deploy wrapper, with a full path to known_hosts based on
    $cluster_dir. So it'll work for anyone using utils/deploy, from any
    directory (as before).

    (b) If ansible_ssh_common_args is empty, roles/common will point
    UserKnownHostsFile to «{{playbook_dir}}/known_hosts". So it'll work
    for anyone using roles/common, even if they don't use utils/deploy.

    (a) We set UserKnownHostsFile to "clusters/$cluster/known_hosts" in
    the inventory. So it'll work for people using utils/ansible from the
    TPA/CustomCloud top-level directory.

In any other situations, ansible_ssh_common_args should be overriden
appropriately to set the correct path.