Always pass in an absolute path to UserKnownHostsFile
We can't be sure what ssh'd cwd will be (we really should be able to,
but I keep running into problems that I don't want to investigate right
now), so the simple answer is to forget about simplified special-case
paths (i.e., ./clusters/x) and use an absolute path based on $utils.

This makes utils/deploy and deploy.yml work always. Or I hope it does.