S3 buckets exist across regions; we don't need one in each region
We don't strictly _need_ to specify the region for the bucket at all,
but I err on the side of assuming that the first region specified in
instances[] is the "principal" region. Accessing objects in our bucket
from some other region will incur some additional latency, that's all.

Gülçin ran into this problem when deploying a cross-region BDR cluster,
and tested the proposed fix.