Disallow the use of 'node' with 'instance_count' in instances[]
Either every instance should explicitly specify its node number, or none
should (in which case instance_count can be used, and will fill in node
numbers).

It may be desirable in future to teach expand_instances to permit some
node numbers to be "pinned" explicitly, in which case this test will
need to become smarter. But we don't do that now, so comparing total
counts is good enough.