Run ssh-add when the key hasn't been registered with ssh-agent
We used to do so only when the key was first generated, but now we'll do
it every time. This handles the case where you've provisioned a cluster
once, and are re-provisioning it after an ssh-agent restart or some such
thing (e.g., you weren't running ssh-agent before, but are this time).

This is something that Nikhil and Richard have both complained about in
the past.