Allow mount options to be specified
The "noauto" in the earlier code crept in from LUKS support, where we
could not mount the device at boot without the passphrase. Instead, we
use 'defaults, noatime' (a reasonable start for ext4 PGDATA); callers
can specify other options, and LUKS (when it's reintegrated) can add a
'noauto' to the end.