Here's where we can install python-psycopg2
Note that the same package name works on Debian/Ubuntu/RedHat.

This is unsatisfying, because if we install Postgres using postgres/src,
we won't be able to use the postgresql_* modules. We could solve that by
installing psycopg2 (using pip) in postgres/src, ugly as that sounds. We
can just ignore the problem for now.