Include based on ansible_os_family
This is really just an illustration: if we don't have to differentiate
between Debian and Ubuntu (which is sometimes true, sometimes not), we
can include {{ ansible_os_family }}.yml and avoid writing an Ubuntu.yml
that does nothing but include (or, worse, duplicate) Debian.yml.

If we find we need to do something Ubuntu-specific in future, we can
change this back to {{ ansible_distribution }}, and still include or
replace Debian.yml as Ubuntu.yml.

This is equally applicable to tasks and vars files.