Get rid of some more Debian/Ubuntu duplication
Switch from ansible_distribution to ansible_os_family for includes.

Also, don't specify defaults for pgversion—it must be defined at the top
level (i.e., deploy.yml) without any opportunities for confusion within
the individual roles. It would not be inappropriate—though excessive in
practice—to assert that it's defined in every role that needs it.