Switch to using defaults/ for postgres_data_dir default values
The problem with using include_vars is that the desired default values
actually override anything that the caller may specify. We don't want
that, so we define low-priority distribution-specific defaults to use
only if the caller hasn't specified anything.

Also, we give up on trying to force the same PGDATA location across
distributions. If someone cares enough to want a specific platform,
they're certainly going to be surprised if their PGDATA doesn't fit
whatever the distribution's default location is.