Make --data-checksums an explicit default initdb option
The old postgres/pkg role treated --data-checksums as a hidden/hardwired
default. Here we make it explicit.

Callers can specify a different set of options. At the moment, there's
no way to augument the defaults, only replace them. Given the specific
nature of initdb options, that's probably not a problem in practice.