Use pg_createcluster if available, but fallback to initdb
Note that the pglogical testing cluster already depended on the initdb
behaviour, so this is not a theoretical concern.

For the moment, I haven't changed the pg_createcluster invocation, but
it's going to change soon enough: neither the port nor the automatic
start are entirely desirable here.

(Aside: it should be possible to install only postgresql-common and use
it with a postgres built from source. We don't currently support that,
but it may become sensible in future.)