Pause after instance profile creation
This is an attempt to address errors of the following kind when creating
an S3 bucket with a policy allowing access to a newly-created principal:

    Invalid principal in policy

Or like this, when launching an EC2 instance:

    InvalidParameterValue: Value (TPA_profile) for parameter iamInstanceProfile.name is invalid. Invalid IAM Instance Profile name
    InvalidParameterValue: IAM Instance Profile \"arn:aws:iam::324077455750:instance-profile/TPA_profile\" has no associated IAM Roles

There is some indication that some AWS regions exhibit this problem more
often than others; however, it's more likely that this is a function of
the provisioner's connectivity to said region.

Unfortunately, this doesn't actually seem to solve the problem. Although
re-running the playbook immediately after "Invalid principal in policy"
does fix the problem, a 60s pause alone did not. Nor did 120s.

I'm committing this just for the record. Will revert shortly.