Split up sys/ssh into sys/ssh/{keygen,install,authorize}
The external interface to sys/ssh is still the same, so no changes will
be needed to any existing deploy.yml. But it is now possible for unusual
use cases to be accommodated by calling the sub-roles directly (e.g., to
allow ssh FROM user postgres TO user barman and vice versa).