Write the assertions, remove the TODO
We now expect instances to refer to each other by name, e.g.:

    instances:
        - node: 1
          tags:
            Name: xxx
            backup: yyy
        - node: 2
          tags:
            upstream: xxx
        - node: 3
          tags:
            Name: yyy
            role: [barman,and,maybe,other,things]

So node 3 is the barman server to which node 1 is backed up, and node 2
is a replica of node 1 (all without fragile numeric references).

Names are optional unless an instance needs to be referenced by another.
Any references will be sanity-checked by the assertions added here. Note
that the references are not yet acted upon, but the upstream references
in particular are the mechanism by which we can support cascading
replication.

(Aside: it boggles the mind that Jinja2 didn't have a 'contains' test
already.)