Tag instances with the cluster_name
Since we're now assigning instances their own names (at least, since we
can), tag:Name={{cluster_name}} is no longer a reliable way to find all
instances. So we make sure every instance is tagged with the Cluster
name as well.

We also allow cluster_tag to be overriden on the command-line. This was
what allowed me to get rid of the untagged cluster instances:

    deprovision test/platforms -e cluster_tag=key_2Q_testplatforms