Use the 'ip_addresses' setting when generating known_hosts
hostvars.yml now sets ip_addresses to ['public_ip','private_ip'], and we
can expand the list (e.g., with vpn_ip) when appropriate and known_hosts
will expand accordingly.

The first part of this commit accidentally sneaked in to an earlier
unrelated commit, but since I pushed it already, it's not worth going
back and fixing it.