Remove pg_hba.conf tweaking altogether
We're just going to trust the private subnet, so we don't need to do
this at all. Also, if we later feel the need to add restrictions, we can
do it differently. For example, on a given primary, pg_hba.conf could be
set up to trust its Barman server (via 'tag_Name_{{ backup }}') and only
its declared replicas—everything can be done on the basis of defined
groups from the inventory based on tags.

That would be a loop in the pg_hba.conf template, and wouldn't need to
be tweaked from the individual barman/repmgr roles at all; that task can
be contained within postgres/config where it belongs anyway.