Adopt a more conservative solution to role variable leakage
Since the effects of set_fact persist beyond the role in which they are
executed (as they should), they are not suitable for variables that are
meant to be used only within a role. Even though we can work around the
problem by overriding ssh_key_file (as I did in an earlier commit), it
remains risky (especially since the sys/ssh/xxx roles are meant to be
used directly in certain circumstances).

Rather than depend on eternal vigilance, we use vars named _xxx to act
as private versions of xxx within a role, and set private_role_vars in
ansible.cfg to restrict their scope (while allowing callers to override
them if desired).