Fix sys/ssh, somewhat by trial and error
The root of the problem was that some (unknown) thing was setting an
ssh_key_file variable (== "id_testplatforms", the key I was using to
authenticate to the instances), which overrode sys/ssh/*'s vars and
uploaded the wrong key etc. Renaming the variable made the problem
disappear.

Separately, I tried to find some way to specify the input vars in an
unambiguous manner that could be passed on to subsidiary roles. I was
not able to find such a way, due to various bugs (summarised in the
comments in sys/ssh/tasks/main.yml).

In light of the earlier name conflict and the fact that "user:" is an
ansible setting sometimes, we rename user: to ssh_user: for these roles.