Add plperl and plv8 by default
By doing this here (i.e., in postgres/pkg), we prevent postgres/src
builds from reusing the same instructions. But that's OK, because we
can't satisfy the postgresql dependencies of OS packages with
postgres/src builds anyway.

postgres/src is for special-case use, and it's assumed that it will be
accompanied by suitable postgresql/ext instructions from the caller.