Make ip_address == private_ip by default
Now when we set up openvpn, we just append openvpn_ip to the
ip_addresses list (which is used to generate known_hosts and so on[1]),
and set ip_address to openvpn_ip, and everything that uses ip_address
will transparently use the right address.

[1] Speaking of which, roles/common/tasks/hosts will have to be executed
later, after openvpn has had a chance to introduce the new address. No
problem, we'll deal with it when we get there.