Fill in barman configuration tasks
We have to install a per-client file in /etc/barman.d (not overwrite
/etc/barman.conf!), and configure a backup cron job. These are most
naturally done by executing on the host to be backed up, and delegating
to their designated backup server, as in client.yml here.

We can use defined defaults/, and derive values in vars/ based on the
host and/or variables specified by the caller. So, for example, when we
do Barman-as-a-service deployments, we can define the string based on
customerid/serviceid in a single place.

(Not yet completed; the dependent roles and templates need editing.)