Add LUKS encryption support
Note that the cryptsetup invocations are highly version specific (e.g.,
RedHat has an older version of cryptsetup that does not accept certain
options), and Haroon and I spent quite some time making sure that they
work well across the distributions we care about. Caution is advised.

This depends on the passphrase being made available as described in
issue #91.