Initial revision of postgres/config
We can no longer leave pg_hba.conf alone by default, so we have to merge
the template with the overriden template for postgres_hba_settings
(which is used by pglogical).

We also move postgres_conf_dir settings to postgres/pkg, and rename
postgres_config_xxx to _conf_xxx, since it's postgresql.conf.

Next, we need to update the templates.