Move _vault_xxx definitions to vars
Here's the idea: if you put _xxx in defaults, then the caller has to
explicitly specify "_xxx" when including the role to override, and the
"private_role_vars" mechanism doesn't apply (to things in defaults). But
we want to provide an overridable default value that doesn't leak (so we
can't use set_fact), so we do the following in vars/main.yml:

    _xxx: "{{ xxx|default('yyy') }}"

Then the caller can do:

    - role: x
    - role: x
      vars:
        xxx: pqr

and _xxx inside our role will be set to 'yyy' and 'pqr' respectively.