Simplify pg_hba.conf handling
The earlier rules led to unnecessarily complex logic about what template
to use, and what lines would be placed where. The new rules are simple:
sensible default, append lines, replace default with lines, or override
the whole thing and do whatever you want.

The default is based on what was in pkg++; it's not sensible yet, but
will be soon.