Replace use of play_hosts with groups[cluster_tag]
The problem with play_hosts is that we sometimes use it to mean "all
hosts in the cluster" (e.g., when generating /etc/hosts), but that's not
what it actually means. It can be different if you run the playbook with
«--limit tag_os_Debian» or otherwise limit the hosts in the play (e.g.,
by using serial).

Instead, we can use groups[cluster_tag], which is guaranteed to mean
"all hosts in the cluster" on any host that evaluates the expression.

It may still be appropriate to use play_hosts in some situations, but
/etc/hosts and /etc/ssh/ssh_known_hosts aren't them.