Minor tweaking of the cluster_vars mechanism
There are three places that can set inventory variables for provisioned
hosts in a cluster:

1. cluster_vars in config.yml
2. platform_vars in provision.yml
3. platforms/common/set-cluster-var.yml

Any variables defined this way will be written to the static inventory
file during provisioning, and thence be available to hosts during the
deployment process.

It is a little strange to treat "customerId" and "instance_profile_arn"
in the same way, but separating them would mean coming up with a good
name for variables generated/set by the provisioning process. That is
hard to do right now because there are so few examples. We can review
this once the mechanism is more widely used. Meanwhile, it does no harm
to make those values available (though it might be a mistake to use any
of those values during deployment).

It may also eventually be necessary to put variables like this into
group_vars/all instead of directly into inventory/00-clustername, so
that they can be overriden on a per-host basis.