Add some default templates
This isn't all of the default configuration we want to generate, but
it's a start (based on earlier work, plus more available overrides).

Note that "fsync=on" can't be overriden just by setting "fsync: off"
when invoking the role. If you really want to do this, you have to put
"fsync: off" into postgres_conf_settings, and it'll be written into the
override file.

Fixes #73