Fill in the remaining postgres conf templates
There are obviously other parameters that we may want to set a default
value for or be able to override; and there are many opportunities for
automatic tuning as well (e.g., effective_cache_size, some of the I/O
cost settings, etc.).

The handling of shared_preload_libraries is as yet insufficient.