Make set-variable work
Unfortunately, when we do a cross-role include, the template resolution
path breaks (this may be a recent change), but tasks/templates/ is one
of the things that seems to both work and be intended (the latter isn't
exactly an endorsement, but it'll do for the moment).

Of course, the real solution is:

    - include_role: name=postgres/config tasks_from=set-variable.yml
      vars:
        name: …
        value: …

but as we've seen already, that just doesn't work yet.