Move certificate generation to postgres/config
Having a key owned by root/ssl-cert is supported on Debian (presumably
through a custom patch), but not on RHEL. So the key/cert we generate
must be owned by postgres:postgres and at least the key must have
permissions 0600 or less.

So we generate it later when we know the postgres user has been created,
and just set the right ownership straightaway.