Fix the "Invalid Principal in Policy" problem once and for all
The fundamental problem was that iam returns an "instance_profile_arn",
which is NOT the same as the ARN for the instance profile role, and not
the right thing to use in the S3 bucket policy. (Yes, that's strange and
annoying.)

It used to work on rerunning the playbook because we wouldn't store the
instance_profile_arn in vars.json (since the s3_bucket task failed), and
the next time we'd use aws to discover the instance profile role's ARN,
which worked. This is why sleeping for long periods didn't work, but
rerunning always did.

Now we just ignore the instance_profile_arn and always use aws to find
the right ARN for the instance profile role, and all is well.

Many thanks to linuxdynasty (Allen Sanabria) for his help to find the
problem.