Move db user creation into a separate include file
This is somewhat WIP; there's a clear repetition with the list of
secrets in roles/common, and it's probably better to generate them
here, where they're needed, rather than having a separate list. But
for the moment, I'm just trying to avoid the repetitive calls to the
postgresql_user module.