Disable certificate usage verification for now
This used to be "ns-cert-type client|server", but the nsCertType
extension is non-standard, and one is supposed to instead use the
extended key usage attributes instead, hence remote-cert-tls.

We do need to do this, it's important. We would need to use a custom
openssl.cnf file with different "-extension section"s for client and
server certificates. easy-rsa does the right thing as a reference.