Set «static: no» on Barman includes
Without this, we break this recently-added test:

    commit fb50698da32f36aaff5359f466679a7b8ff3f80b
    Author: cinerama70 <cinerama70@users.noreply.github.com>
    Date:   Wed Sep 28 16:24:32 2016 -0700

        Graceful error handling when delegate_to host null (#17453) (#17783)

        Currently, if the host specified in delegate_to for a task is null,
        Ansible will crash with a stack trace. Add a check for this state
        and handle the error appropriately.

What happens is that even though we expect the task to _run_ only for
hosts with backup != '' (the include condition), we still try to
evaluate the delegation for the other hosts, and raise an error.

After subsequent discussion with abadger and cinerama, we're changing
the test upstream to avoid this problem, but even so, this change
expresses our intention more clearly and incidentally cuts down on
"skipped" output too.