Use separate files for each backup cronjob
We delegate the cron task to the backup server from multiple other hosts
simultaneously, which causes a race condition in changing the file. It's
possible to run the tasks successfully to completion and still end up
with only one server's backup cronjob in the file (it may even remove
the default package-added "barman cron" job somehow).

If we use separate files, the problem is eliminated, along with a
potential upgrade conflict between a customised /etc/cron.d/barman and
the upstream packaged file.