Instead of building up a list of secrets, generate them when needed
Now any role that needs a secret should declare a dependency like this:

    - role: secret
      vars:
        secret_name: xyzzy_password

This will make the secret be generated and encrypted and loaded, all in
one step. No need for explicit include_vars any more.

The new pgbouncer role needed a password, but didn't change the list in
the common role (and so would have ended up breaking), but we're now
past the threshold where we can keep a list in sync.

Note the unfortunate change where we have to remove a "vars" from a
dependency invocation of postgres/createuser. It fails only for the
repmgr/final role (maybe because of nesting level? not sure).