Tighten up and clarify the barman role in various ways
Using initialise_backup to transfer state from the barman role to the
barman/final role wasn't a good idea. If everything worked, it was fine,
but if it didn't, then rerunning the playbook wouldn't re-initialise an
earlier failed backup. It's one thing when you use the flag to say "this
role has been run" (cf. initialise_replication from repmgr) but not such
a good idea when it's an "action" flag that can't be inferred again.

We change barman/final to split it into server/client (which fixes the
awkward pgpass-generation problem straightaway) and make it not depend
on initialise_backup. We still use it in postgres/final/first-backup but
that's just a matter of figuring out how to check if there's any backup
for a given server already.

We also try to replace the awkward "we hope this is long enough" sleeps
in first-backup with more specific tests (is pg_receivexlog running? has
it seen the segment switch?). It's not quite right yet (the latter test
needs to do something more than check that flush_location has advanced),
but it's better than what we had earlier.

Various other minor fixes besides.