Don't define localhost explicitly in inventory/static
Somewhere during the Ansible 2.0 development cycle, the implicit
localhost entry in the inventory broke, and I added this explicit entry
to workaround that problem. The problem was later fixed, but an exciting
new problem was introduced—that having an explicit localhost entry would
suppress the use of sys.executable as ansible_python_interpreter for
local connections.

Since I'm not a fan of specifying «-e ansible_python_interpreter=…» for
every utils/ansible invocation, and since the original problem is gone,
I removed the localhost entry, only to find ansible spewing warnings…
about the implicit localhost being the only available entry.

So we don't want to omit the host file entirely (warning), nor an empty
host file (warning), and we don't want localhost in it (wrong python).
But we can have unicode hostnames to express how much we love this
situation, so why not.