Change the initall logic. Instead of creating a grand pgxc_ctl.conf we now add components piecemeal. Each instance gets a role: 'gtm', 'gtm-standby', 'coordinator', 'datanode', 'datanode-replica' now.
Additionally, we EXPLICITLY mention the index of each coordinator/
datanode/replica that resides on each instance based on the above
roles.

We use 'dn_list' and 'dn_replica_list' to indicate indexes of
datanodes and its replicas respectively

We use 'co_list' and [optionally] 'co_replica_list' for coordinators
as well.