Add an upstream_root filter plugin
Until now, the 'upstream' of a given node would either be empty (if it
was not a replica) or pointed directly to the corresponding primary. To
support cascading replication, we need to be able to point the upstream
to a replica. We need to do the setup in one pass, however, so we can't
wait for intermediate replicas to be set up and clone from them to set
up further replicas.

So we need to find the root of the cascading setup, which is what this
filter does. Now "{{ inventory_hostname|upstream_root(hostvars) }}" will
return the ancestral primary for any node.

This can't be done using a set_fact/until loop in ansible because if the
until condition fails, the inner set_fact's results are not propagated,
i.e., the following invocation would just keep retrying the set_fact
with the same arguments.

    - set_fact:
        upstream_root: "{{ upstream }}"
      until: hostvars[upstream_root].upstream == ''